+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Add an asset.logs service for retrieving logs that reference an asset #850](https://github.com/farmOS/farmOS/pull/850)
 - [Add farmOS-timeline library (experimental) #862](https://github.com/farmOS/farmOS/pull/862)
 - [Build multi-platform Docker images/manifests #864](https://github.com/farmOS/farmOS/pull/864)
+- [Add a Soil texture field to Lab test logs #873](https://github.com/farmOS/farmOS/pull/873)
 
 ### Changed